backend/requirements.txt
fastapi==0.129.2
uvicorn[standard]==0.41.0
python-multipart==0.0.22
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
asyncpg==0.30.0
pydantic-settings==2.12.0